faster way to walk a directory tree
Today I noticed high CPU usage when using pollen inside a VM. Though
this is normal when doing polling I'd like to reduce it to a minimum and
this is a start.

The filepath.Walk function in the Go standard library does an lstat to
get the FileInfo for each entry found which we don't require. The
alternative presented here seems to work for Linux but OS X (and Windows
if we start caring) needs more testing.