Improve reloading on vagrant vboxfs/NFS
Detecting changes based on files being modified in the last X seconds
does not work well on NFS as there seems to be an arbitrary delay in
receiving mtime changes made from the host. This change works around
that by caching all the last mtimes of the files we have seen and
compare against that.